Feat: New flippers options
Attribute 1: Flipper dive time
Attribute 2: Flipper cooldown time
Flag 2: Cancellable dive
Updated V_ITEMS to 45
Replaced various `diveclk>30` calls with `isDiving()` calls, which previously just returned `diveclk>30`, but now handles the new options.
Included sanity checks for this at several points.
If no flippers are had, the default values of 50 (dive) and 30 (cooldown) will be used.

+++ src/ffscript.cpp
@@ -16290,7 +16290,7 @@ bool FFScript::warp_link(int warpType, int dmapID, int scrID, int warpDestX, int
 		{
 			//Z_scripterrlog("FFCore.warp_link reached line: %d \n", 15936);
 			bool wasswimming = (Link.getAction()==swimming);
-			byte olddiveclk = Link.diveclk;
+			int olddiveclk = Link.diveclk;
 			ALLOFF();
 			if ( !(warpFlags&warpFlagDONTKILLMUSIC) ) music_stop();
 			if ( !(warpFlags&warpFlagDONTKILLSOUNDS) ) kill_sfx();
